+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [1.2.0]
+
+* Added new `SelectionAlgorithm::Random`.
+* Added new `SelectionAlgorithm::Stochastic`.
+* `SelectionAlgorithm` now derive `PartialEq`, `Eq`, `PartialOrd` and `Ord`.
+* Added unit testing for `Selection` trait.
+
 # [1.1.0]
 
 * Added benchmarks. You can run benchmarks using `cargo bench`. The benches are all inside `benches/`.